+++ packages/core/src/index.ts
@@ -45,5 +45,6 @@ export * from "./setup";
 export * from "./tachoTime";
 export * from "./vehicleCosts";
 export * from "./vehiclePnl";
+export * from "./weeklyRest";
 export * from "./workTime";
 export * from "./zodErrors";